+++ update_finish.sh
@@ -59,11 +59,12 @@ popd
 # make sure we are on node 8.x.x
 sudo /home/pi/sisbot-server/sisbot/node_update.sh > /var/log/sisyphus/node_update.log
 
+# make sure python is installed
+sudo /home/pi/sisbot-server/sisbot/install_python.sh
+
 echo "Upgrade_Finish completed"
 
 # 1.0-1.2 reboot necessity, to make sure bluetooth updates self
 if [ -z "$1" ]; then
 	sudo reboot
 fi
-
-