+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - CSON RGBW Offset
   - Homing offset after successful sensored home
   - Clean Log files at midnight each day
+  - Auto-install python requirements
 
 # 2019-08-26 1.10.41
   - Light patterns